--- a/pytrip/lib/core.c
+++ b/pytrip/lib/core.c
@@ -862,7 +862,6 @@ static PyObject* binary_search_intersection(PyObject *self, PyObject *args){
                 current_direction = (x_0 < x_1);
                 while (l <= r){
                     m = (l+r)/2;
-                    printf("%d ", m);
                     x_0 = *((double*)PyArray_GETPTR2(vec_slice, m, 0));
                     x_1 = *((double*)PyArray_GETPTR2(vec_slice, m+1, 0));
                     if((x_0 >= depth && x_1 < depth) || (x_1 >= depth && x_0 < depth)){
@@ -882,7 +881,6 @@ static PyObject* binary_search_intersection(PyObject *self, PyObject *args){
                         }
                     }
                 }
-                printf("\n");
 
                 y_0 = *((double*)PyArray_GETPTR2(vec_slice, m, 1));
                 y_1 = *((double*)PyArray_GETPTR2(vec_slice, m+1, 1));
@@ -907,7 +905,6 @@ static PyObject* binary_search_intersection(PyObject *self, PyObject *args){
                 current_direction = (y_0 < y_1);
                 while (l <= r){
                     m = (l+r)/2;
-                    printf("%d ", m);
                     y_0 = *((double*)PyArray_GETPTR2(vec_slice, m, 1));
                     y_1 = *((double*)PyArray_GETPTR2(vec_slice, m+1, 1));
                     if((y_0 >= depth && y_1 < depth) || (y_1 >= depth && y_0 < depth)){
@@ -927,8 +924,6 @@ static PyObject* binary_search_intersection(PyObject *self, PyObject *args){
                         }
                     }
                 }
-
-                printf("\n");
 
                 x_0 = *((double*)PyArray_GETPTR2(vec_slice, m, 0));
                 x_1 = *((double*)PyArray_GETPTR2(vec_slice, m+1, 0));