how it was:
```External
sagittal time: 17.839640617370605 s
coronal time: 12.245706558227539 s
External elapsed: 30.128231048583984
CTV_6000
sagittal time: 0.4264366626739502 s
coronal time: 0.5313541889190674 s
CTV_6000 elapsed: 0.9587903022766113
BrainStem
sagittal time: 0.14860320091247559 s
coronal time: 0.1306602954864502 s
BrainStem elapsed: 0.2802610397338867
Chiasm
sagittal time: 0.027945995330810547 s
coronal time: 0.016968965530395508 s
Chiasm elapsed: 0.044914960861206055
LacrimalGland_L
sagittal time: 0.015944957733154297 s
coronal time: 0.02394270896911621 s
LacrimalGland_L elapsed: 0.03988766670227051
LacrimalGland_R
sagittal time: 0.01400899887084961 s
coronal time: 0.020938634872436523 s
LacrimalGland_R elapsed: 0.03494763374328613
Lens_L
sagittal time: 0.010988712310791016 s
coronal time: 0.010943889617919922 s
Lens_L elapsed: 0.021932601928710938
Lens_R
sagittal time: 0.011013507843017578 s
coronal time: 0.008932352066040039 s
Lens_R elapsed: 0.019945859909057617
OpticNerve_L
sagittal time: 0.029902935028076172 s
coronal time: 0.05588555335998535 s
OpticNerve_L elapsed: 0.08578848838806152
OpticNerve_R
sagittal time: 0.04787397384643555 s
coronal time: 0.054892539978027344 s
OpticNerve_R elapsed: 0.10276651382446289
CTV_5000
sagittal time: 0.9807066917419434 s
coronal time: 1.2633070945739746 s
CTV_5000 elapsed: 2.2480032444000244
Brain
sagittal time: 2.2313084602355957 s
coronal time: 3.0970230102539062 s
Brain elapsed: 5.338260650634766
Spinalcord
sagittal time: 0.12816548347473145 s
coronal time: 0.14362144470214844 s
Spinalcord elapsed: 0.27375149726867676
total voi contours calc time: 39.57941555976868 s
```

now:
```
voi: External, elapsed 4.589036226272583 s
voi: CTV_6000, elapsed 0.377030611038208 s
voi: BrainStem, elapsed 0.11072230339050293 s
voi: Chiasm, elapsed 0.03294968605041504 s
voi: LacrimalGland_L, elapsed 0.02789473533630371 s
voi: LacrimalGland_R, elapsed 0.02592182159423828 s
voi: Lens_L, elapsed 0.013965129852294922 s
voi: Lens_R, elapsed 0.01599907875061035 s
voi: OpticNerve_L, elapsed 0.07480025291442871 s
voi: OpticNerve_R, elapsed 0.07181715965270996 s
voi: CTV_5000, elapsed 0.5401036739349365 s
voi: Brain, elapsed 0.9055707454681396 s
voi: Spinalcord, elapsed 0.09624505043029785 s
total elapsed 6.892634630203247 s
```