@rozchlastywacz 2022-01-10
```
binary search

voi: External, elapsed 3.2484934329986572 s
voi: Bladder, elapsed 0.34410786628723145 s
voi: PenileBulb, elapsed 0.034906625747680664 s
voi: FemoralHead_L, elapsed 0.14261770248413086 s
voi: FemoralHead_R, elapsed 0.2493736743927002 s
voi: Rectum, elapsed 0.21650290489196777 s
voi: iCTVp_04_6000, elapsed 0.14586639404296875 s
voi: Rectalwall, elapsed 0.30973076820373535 s
voi: iCTVp+sb_4800, elapsed 0.19720792770385742 s
voi: Ref, elapsed 0.0030045509338378906 s
total elapsed 4.9147093296051025 s

voi: duodenum, elapsed 0.6756901741027832 s
voi: kidney_L, elapsed 0.5653185844421387 s
voi: kidney_R, elapsed 0.5904614925384521 s
voi: liver, elapsed 1.673318862915039 s
voi: pancreas, elapsed 0.4058980941772461 s
voi: spleen, elapsed 0.5664994716644287 s
voi: Spinalcord, elapsed 0.29897093772888184 s
voi: stomach, elapsed 0.703667402267456 s
voi: iCTV_5040, elapsed 0.38197851181030273 s
voi: iCTV_5580, elapsed 0.2522592544555664 s
voi: Ref, elapsed 0.0030291080474853516 s
voi: External, elapsed 3.976346969604492 s
total elapsed 10.102450132369995 s
```
```
linear search

voi: External, elapsed 4.004110097885132 s
voi: Bladder, elapsed 0.3600771427154541 s
voi: PenileBulb, elapsed 0.034906864166259766 s
voi: FemoralHead_L, elapsed 0.14262127876281738 s
voi: FemoralHead_R, elapsed 0.14759278297424316 s
voi: Rectum, elapsed 0.20450735092163086 s
voi: iCTVp_04_6000, elapsed 0.13564300537109375 s
voi: Rectalwall, elapsed 0.4699103832244873 s
voi: iCTVp+sb_4800, elapsed 0.1924901008605957 s
voi: Ref, elapsed 0.0019555091857910156 s
total elapsed 5.69935941696167 s

voi: duodenum, elapsed 0.6545774936676025 s
voi: kidney_L, elapsed 0.5595252513885498 s
voi: kidney_R, elapsed 0.5341627597808838 s
voi: liver, elapsed 1.7050607204437256 s
voi: pancreas, elapsed 0.4287686347961426 s
voi: spleen, elapsed 0.40099000930786133 s
voi: Spinalcord, elapsed 0.514641523361206 s
voi: stomach, elapsed 0.56947922706604 s
voi: iCTV_5040, elapsed 0.2593080997467041 s
voi: iCTV_5580, elapsed 0.21346664428710938 s
voi: Ref, elapsed 0.002001523971557617 s
voi: External, elapsed 4.479854345321655 s
total elapsed 10.331799507141113 s
```
```binary if len(contour) > 50 else linear

voi: External, elapsed 3.6747403144836426 s
voi: Bladder, elapsed 0.49168825149536133 s
voi: PenileBulb, elapsed 0.06383132934570312 s
voi: FemoralHead_L, elapsed 0.22048211097717285 s
voi: FemoralHead_R, elapsed 0.15608906745910645 s
voi: Rectum, elapsed 0.23301434516906738 s
voi: iCTVp_04_6000, elapsed 0.1396348476409912 s
voi: Rectalwall, elapsed 0.29243040084838867 s
voi: iCTVp+sb_4800, elapsed 0.18539118766784668 s
voi: Ref, elapsed 0.001993894577026367 s
total elapsed 5.484140157699585 s

voi: duodenum, elapsed 0.5655210018157959 s
voi: kidney_L, elapsed 0.43078017234802246 s
voi: kidney_R, elapsed 0.4647223949432373 s
voi: liver, elapsed 1.3007078170776367 s
voi: pancreas, elapsed 0.24189496040344238 s
voi: spleen, elapsed 0.26429104804992676 s
voi: Spinalcord, elapsed 0.24135303497314453 s
voi: stomach, elapsed 0.5096778869628906 s
voi: iCTV_5040, elapsed 0.2658402919769287 s
voi: iCTV_5580, elapsed 0.23139572143554688 s
voi: Ref, elapsed 0.001997709274291992 s
voi: External, elapsed 3.9494316577911377 s
total elapsed 8.474636316299438 s
```