@rozchlastywacz 2022-01-11
![image](https://user-images.githubusercontent.com/44035398/148994023-2666674d-54bb-405f-a6b6-7998a725ffdd.png)

same test launched twice, on the same machine, same app, nothing changed, except for order of loaded files:

first launch:
```
binary if len(contour) > 50 else linear
if binary:
  binary if binary_search_range > 50 else linear

voi: External, elapsed 3.1879823207855225 s
voi: Bladder, elapsed 0.4182724952697754 s
voi: PenileBulb, elapsed 0.05209183692932129 s
voi: FemoralHead_L, elapsed 0.18416857719421387 s
voi: FemoralHead_R, elapsed 0.13272380828857422 s
voi: Rectum, elapsed 0.18000411987304688 s
voi: iCTVp_04_6000, elapsed 0.10998654365539551 s
voi: Rectalwall, elapsed 0.24477267265319824 s
voi: iCTVp+sb_4800, elapsed 0.15303611755371094 s
voi: Ref, elapsed 0.0 s
total elapsed 4.710136413574219 s

voi: duodenum, elapsed 0.3991119861602783 s
voi: kidney_L, elapsed 0.40128469467163086 s
voi: kidney_R, elapsed 0.40085625648498535 s
voi: liver, elapsed 1.3769521713256836 s
voi: pancreas, elapsed 0.30296897888183594 s
voi: spleen, elapsed 0.31831812858581543 s
voi: Spinalcord, elapsed 0.25022435188293457 s
voi: stomach, elapsed 0.5239012241363525 s
voi: iCTV_5040, elapsed 0.27240800857543945 s
voi: iCTV_5580, elapsed 0.20270037651062012 s
voi: Ref, elapsed 0.0 s
voi: External, elapsed 4.1474363803863525 s
total elapsed 8.611509084701538 s
```
second launch, faster one:
```
binary if len(contour) > 50 else linear
if binary:
  binary if binary_search_range > 50 else linear

voi: External, elapsed 2.5636441707611084 s
voi: Bladder, elapsed 0.2939891815185547 s
voi: PenileBulb, elapsed 0.030646562576293945 s
voi: FemoralHead_L, elapsed 0.12221169471740723 s
voi: FemoralHead_R, elapsed 0.12476181983947754 s
voi: Rectum, elapsed 0.16330504417419434 s
voi: iCTVp_04_6000, elapsed 0.1119847297668457 s
voi: Rectalwall, elapsed 0.25495123863220215 s
voi: iCTVp+sb_4800, elapsed 0.16068315505981445 s
voi: Ref, elapsed 0.0 s
total elapsed 3.8270554542541504 s

voi: duodenum, elapsed 0.3780391216278076 s
voi: kidney_L, elapsed 0.3394308090209961 s
voi: kidney_R, elapsed 0.3353567123413086 s
voi: liver, elapsed 1.3653676509857178 s
voi: pancreas, elapsed 0.2877640724182129 s
voi: spleen, elapsed 0.3064582347869873 s
voi: Spinalcord, elapsed 0.2019953727722168 s
voi: stomach, elapsed 0.41782331466674805 s
voi: iCTV_5040, elapsed 0.22156357765197754 s
voi: iCTV_5580, elapsed 0.17237448692321777 s
voi: Ref, elapsed 0.0 s
voi: External, elapsed 3.6047849655151367 s
total elapsed 7.6392621994018555 s
```